+++ spec/taro/types/shared/fields_spec.rb
@@ -15,7 +15,13 @@ describe Taro::Types::Shared::Fields do
   end
 
   it 'raises when evaluating fields without type' do
-    expect { example.field :bar, null: true }.to raise_error(Taro::ArgumentError, /type/)
+    expect { example.field :bar, null: true }
+      .to raise_error(Taro::ArgumentError, /type.*must be given/)
+  end
+
+  it 'raises when evaluating fields with non-string type' do
+    expect { example.field :bar, type: S::StringType, null: true }
+      .to raise_error(Taro::ArgumentError, /type must be a String/)
   end
 
   it 'raises when evaluating fields without null' do
@@ -24,25 +30,27 @@ describe Taro::Types::Shared::Fields do
 
   it 'raises when redefining fields' do
     example.field :bar, type: 'String', null: true
-    expect { example.field :bar, type: 'Boolean', null: false }.to raise_error(Taro::Error, /defined/)
+    expect { example.field :bar, type: 'Boolean', null: false }
+      .to raise_error(Taro::Error, /previously defined/)
   end
 
   it 'takes array_of instead of type' do
     example.field :bar, array_of: 'String', null: true
     field = example.fields[:bar]
-    expect(field.type).to be < Taro::Types::ListType
+    expect(field.type).to be < T::ListType
   end
 
-  # it 'handles array_of with nested types' do
-  #   example.field :bar, array_of: 'Object', null: true
-  #   field = example.fields[:bar]
-  #   expect(field.type).to be < Taro::Types::ListType
-  #   expect(field.type.item_type).to be < Taro::Types::ObjectTypes::ObjectType
-  # end
+  it 'handles array_of with nested types' do
+    stub_const('FooType', Class.new(T::ObjectType))
+    example.field :bar, array_of: 'FooType', null: true
+    field = example.fields[:bar]
+    expect(field.type).to be < T::ListType
+    expect(field.type.item_type).to eq FooType
+  end
 
   it 'takes page_of instead of type' do
     example.field :bar, page_of: 'Integer', null: true
     field = example.fields[:bar]
-    expect(field.type).to be < Taro::Types::ObjectTypes::PageType
+    expect(field.type).to be < T::ObjectTypes::PageType
   end
 end